data_preprocessing/requirements.txt
h5py==3.8.0
matplotlib==3.6.3
more_itertools==10.3.0
numpy==1.23.5
pandas==1.5.3
Pillow==10.4.0
py_feat==0.5.1
scikit_image==0.19.3
scipy==1.10.0
tqdm==4.64.1
